pyproject.toml
[tool.poetry]
name = "pwmanager"
version = "0.1.0"
description = "Password manager for my family"
authors = ["Alen Gozy"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
psycopg2-binary = "^2.9.9"
django-environ = "^0.11.2"
django-crispy-forms = "^2.0"
crispy-bootstrap4 = "^2022.1"
django-rest-framework = "^0.1.0"
drf-spectacular = "^0.26.5"
djangorestframework-simplejwt = "^5.3.0"
django-encrypted-model-fields = "^0.6.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
